rename status util to validatorStatus
