+++ scripts/cc-cli/src/utils/validatorStatus.ts
